Refactor to a server/client architecture
WHAT: First step to later introduce an online/multiplayer version of the game
