--- index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="author" content="digitsensitive" />
-    <title>Snake Legacy</title>
-    <script src="dist/bundle.js"></script>
-  </head>
-  <body>
-    <div id="game"></div>
-  </body>
-</html>
