+++ package.json
@@ -7,6 +7,8 @@
     "Snake Game"
   ],
   "devDependencies": {
+    "@types/express": "^4.17.21",
+    "copy-webpack-plugin": "^11.0.0",
     "expose-loader": "4.0.0",
     "ts-loader": "9.3.1",
     "typescript": "4.8.0-dev.20220730",
@@ -15,11 +17,16 @@
     "webpack-dev-server": "4.9.3"
   },
   "dependencies": {
+    "app-root-path": "^3.1.0",
+    "express": "^4.18.2",
     "phaser": "3.60.0-beta.9",
     "prettier": "2.8.8"
   },
   "scripts": {
-    "dev": "webpack --mode development && webpack serve --mode development"
+    "build:server": "tsc -p tsconfig.json",
+    "build:client": "webpack --config webpack.client.config.js",
+    "build": "npm run build:server && npm run build:client",
+    "start": "node dist-server/server/app.js"
   },
   "repository": {
     "type": "git",
