+++ public/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="author" content="digitsensitive" />
+  <title>Snake Legacy</title>
+  <script src="bundle.js"></script>
+</head>
+
+<body>
+  <div id="game"></div>
+</body>
+
+</html>
