+++ src/client/game.ts
@@ -1,0 +1,12 @@
+import 'phaser';
+import { GameConfig } from './phaser/config';
+
+export class Game extends Phaser.Game {
+  constructor(config: Phaser.Types.Core.GameConfig) {
+    super(config);
+  }
+}
+
+window.addEventListener('load', () => {
+  const game = new Game(GameConfig);
+});
