+++ src/client/phaser/scenes/boot-scene.ts
