+++ src/client/phaser/services/paint-borders.ts
