+++ src/client/phaser/services/position-to-graphics.ts
@@ -1,0 +1,9 @@
+import { Position } from '../../../domain/value-objects/position';
+
+export const positionToGraphics = (position: Position, dotSize: number) => {
+  const initialBorder = dotSize;
+  return {
+    x: position.x * dotSize + initialBorder,
+    y: position.y * dotSize + initialBorder
+  };
+};
