+++ src/domain/components/interfaces/graphics/apple.ts
@@ -1,0 +1,6 @@
+import { Apple } from "../../../objects/apple"
+
+export interface IAppleGraphics {
+  draw(snake: Apple): void
+  update(snake: Apple): void
+}
