+++ src/domain/components/interfaces/graphics/snake.ts
@@ -1,0 +1,6 @@
+import { Snake } from "../../../objects/snake"
+
+export interface ISnakeGraphics {
+  draw(snake: Snake): void
+  update(snake: Snake): void
+}
