+++ src/domain/components/interfaces/input/snake.ts
@@ -1,0 +1,5 @@
+import { Snake } from "../../../objects/snake"
+
+export interface ISnakeInput {
+  update(snake: Snake): void
+}
