+++ src/domain/services/calculate-player-direction.ts
