+++ src/domain/services/is-apple-snake-collision.ts
