+++ src/domain/services/is-border-player-collision.ts
