+++ src/domain/services/is-snake-self-collision.ts
