+++ src/domain/services/is-snake-snake-collision.ts
