+++ src/domain/value-objects/grid-config.ts
