+++ src/domain/value-objects/snake-cursor-keys.ts
@@ -1,0 +1,10 @@
+export type SnakeCursorKeys = {
+  up: Key;
+  down: Key;
+  left: Key;
+  right: Key;
+};
+
+type Key = {
+  isDown: Boolean
+}
