+++ src/domain/value-objects/world.ts
