--- src/game.ts
@@ -1,12 +1,0 @@
-import 'phaser';
-import { GameConfig } from './config';
-
-export class Game extends Phaser.Game {
-  constructor(config: Phaser.Types.Core.GameConfig) {
-    super(config);
-  }
-}
-
-window.addEventListener('load', () => {
-  const game = new Game(GameConfig);
-});
