--- src/value-objects/snake-cursor-keys.ts
@@ -1,6 +1,0 @@
-export type SnakeCursorKeys = {
-  up: Phaser.Input.Keyboard.Key;
-  down: Phaser.Input.Keyboard.Key;
-  left: Phaser.Input.Keyboard.Key;
-  right: Phaser.Input.Keyboard.Key;
-};
