+++ webpack.client.config.js
@@ -1,0 +1,35 @@
+const path = require('path');
+const CopyPlugin = require('copy-webpack-plugin');
+
+module.exports = {
+  entry: './src/client/game.ts',
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'bundle.js'
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        include: path.resolve(__dirname, 'src'),
+        loader: 'ts-loader',
+        options: { configFile: 'tsconfig.client.json' }
+      },
+      {
+        test: require.resolve('phaser'),
+        loader: 'expose-loader',
+        options: { exposes: { globalName: 'Phaser', override: true } }
+      }
+    ]
+  },
+  plugins: [
+    new CopyPlugin({
+      patterns: [
+        { from: 'public', to: '.' },
+      ],
+    }),
+  ],
+  resolve: {
+    extensions: ['.ts', '.js']
+  }
+};
