--- webpack.config.js
@@ -1,32 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  entry: './src/game.ts',
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-    filename: 'bundle.js'
-  },
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        include: path.resolve(__dirname, 'src'),
-        loader: 'ts-loader'
-      },
-      {
-        test: require.resolve('phaser'),
-        loader: 'expose-loader',
-        options: { exposes: { globalName: 'Phaser', override: true } }
-      }
-    ]
-  },
-  devServer: {
-    static: path.resolve(__dirname, './'),
-    host: 'localhost',
-    port: 8080,
-    open: false
-  },
-  resolve: {
-    extensions: ['.ts', '.js']
-  }
-};
